requirements.txt
# Beatify - Python Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Development Tools (Optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0